nit: check if file exists and use provided mime type for extension

+++ lib/mic_web/controllers/whatsapp_controller.ex
@@ -429,32 +429,37 @@ defmodule MicWeb.WhatsAppController do
 
     # Extract the media ID from the audio message
     media_id = audio["id"]
+    mime_type = audio["mime_type"]
 
     # Retrieve the media URL from the WhatsApp API
     media_url = get_media_url(media_id)
+    extension = get_audio_extension(mime_type)
 
     case media_url do
       {:ok, url} ->
         # Download the media file
-        file_path = download_audio_media(url, media_id)
+        file_path = download_audio_media(url, media_id, extension)
 
         if file_path != nil do
           Logger.info("File path: #{file_path}")
 
-          case transcribe_voice_note(file_path, conn, from_number) do
-            {:ok, temporary_reply} ->
-              send_audio_message(conn, from_number, temporary_reply)
+          if File.exists?(file_path) do
+            case transcribe_voice_note(file_path, conn, from_number) do
+              {:ok, temporary_reply} ->
+                send_audio_message(conn, from_number, temporary_reply)
+                File.rm(file_path)
 
-              File.rm(file_path)
+              {:error, reason} ->
+                Logger.error("Transcription Error: #{inspect(reason)}")
+                File.rm(file_path)
+            end
+          else
+            Logger.error("File not found: #{file_path}")
 
-            {:error, reason} ->
-              Logger.error("Transcription Error: #{inspect(reason)}")
-              File.rm(file_path)
-
-              send_message(
-                from_number,
-                error_message
-              )
+            send_message(
+              from_number,
+              "Sorry, an error occurred while processing your audio message."
+            )
           end
         else
           File.rm(file_path)
@@ -500,7 +505,7 @@ defmodule MicWeb.WhatsAppController do
     end
   end
 
-  defp download_audio_media(url, media_id) do
+  defp download_audio_media(url, media_id, extension) do
     config = Application.get_env(:mic, :whatsapp)
     access_token = config[:whatsapp_temporary_access_token]
 
@@ -513,10 +518,17 @@ defmodule MicWeb.WhatsAppController do
         # Save the downloaded media file
         # TODO: ALL OF THE INSTANCES WHERE I WRITE A FILE, NED TO MAKE SURE I DELETE IT IN CASE OF ERROR. IS THERE A BETTER WAY TO DO THIS?
         # TODO: will it always be .ogg? mae sure to save this and mp3 to enum or vars.
-        {:ok, file_path} = Temp.path(%{prefix: "#{media_id}", suffix: ".ogg"})
-        File.write!(file_path, body)
-        Logger.info("Audio downloaded successfully: #{media_id}.ogg")
-        file_path
+        if extension do
+          {:ok, file_path} = Temp.path(%{prefix: "#{media_id}", suffix: ".#{extension}"})
+          File.write!(file_path, body)
+          Logger.info("Audio downloaded successfully: #{media_id}.#{extension}")
+          file_path
+        else
+          {:ok, file_path} = Temp.path(%{prefix: "#{media_id}", suffix: ".ogg"})
+          File.write!(file_path, body)
+          Logger.info("Audio downloaded successfully: #{media_id}.ogg")
+          file_path
+        end
 
       {:ok, %HTTPoison.Response{status_code: status_code}} ->
         Logger.error("Failed to download media. Status code: #{status_code}")
@@ -558,6 +570,15 @@ defmodule MicWeb.WhatsAppController do
       {:error, %HTTPoison.Error{reason: reason}} ->
         Logger.error("Failed to download media: #{inspect(reason)}")
         nil
+    end
+  end
+
+  defp get_audio_extension(mime_type) do
+    case mime_type do
+      "audio/ogg" -> "ogg"
+      "audio/mpeg" -> "mp3"
+      "audio/wav" -> "wav"
+      _ -> nil
     end
   end
